requirements.txt
lunarcalendar
uvicorn 
fastapi 
ephem
pandas
tqdm
numpy
yfinance
yahooquery
scikit-learn
Flask
# request
jsonify
google-cloud-bigquery
python-dotenv
db_dtypes
ib_insync